reorganized fr-config style command files added service provider export command
