+++ src/configManagerOps/FrConfigSecretOps.ts
@@ -1,0 +1,93 @@
+import { frodo } from '@rockcarver/frodo-lib';
+import {
+  createProgressIndicator,
+  printError,
+  stopProgressIndicator,
+  updateProgressIndicator,
+} from '../utils/Console';
+import { SecretSkeleton } from '@rockcarver/frodo-lib/types/api/cloud/SecretsApi';
+import { SecretsExportInterface } from '@rockcarver/frodo-lib/types/ops/cloud/SecretsOps';
+
+const { getFilePath, saveJsonToFile } = frodo.utils;
+const {
+  readSecrets,
+  createSecret: _createSecret,
+  exportSecret,
+  createVersionOfSecret: _createVersionOfSecret,
+  deleteSecret: _deleteSecret,
+  deleteVersionOfSecret: _deleteVersionOfSecret,
+} = frodo.cloud.secret;
+
+/**
+ * Export all secrets to individual files in fr-config-manager format
+ * @param {boolean} includeMeta true to include metadata, false otherwise. Default: true
+ * @param {boolean} includeActiveValues include active value of secret (default: false)
+ * @param {string} target Host URL of target environment to encrypt secret value for
+ * @returns {Promise<boolean>} true if successful, false otherwise
+ */
+type FrConfigSecret = SecretSkeleton & {
+  valueBase64: string;
+};
+async function getFrConfigSecrets(): Promise<FrConfigSecret[]> {
+  const originalSecrets = await readSecrets();
+  return originalSecrets.map((secret) => ({
+    ...secret,
+    valueBase64: `\${${secret._id.toUpperCase().replace(/-/g, '_')}}`,
+  }));
+}
+export async function configManagerExportSecrets(
+  target?: string
+): Promise<boolean> {
+  let secrets: FrConfigSecret[] = [];
+  const spinnerId = createProgressIndicator(
+    'indeterminate',
+    0,
+    `Reading secrets...`
+  );
+  try {
+    secrets = await getFrConfigSecrets();
+    secrets.sort((a, b) => a._id.localeCompare(b._id));
+    stopProgressIndicator(
+      spinnerId,
+      `Successfully read ${secrets.length} secrets.`,
+      'success'
+    );
+    const indicatorId = createProgressIndicator(
+      'determinate',
+      secrets.length,
+      'Exporting secrets'
+    );
+    for (const secret of secrets) {
+      const exportData: SecretsExportInterface = await exportSecret(
+        secret._id,
+        false,
+        target
+      );
+      const [secretKey] = Object.keys(exportData.secret);
+      const fullSecret = exportData.secret[secretKey] as FrConfigSecret;
+      const cleanSecret = {
+        _id: fullSecret._id,
+        description: fullSecret.description,
+        encoding: fullSecret.encoding,
+        useInPlaceholders: fullSecret.useInPlaceholders,
+        valueBase64: `\${${secret._id.toUpperCase().replace(/-/g, '_')}}`,
+      };
+      saveJsonToFile(
+        cleanSecret,
+        getFilePath(`esvs/secrets/${secret._id}.json`, true),
+        false
+      );
+      updateProgressIndicator(indicatorId, `Exported secret ${secret._id}`);
+    }
+    stopProgressIndicator(indicatorId, `${secrets.length} secrets exported.`);
+    return true;
+  } catch (error) {
+    stopProgressIndicator(
+      spinnerId,
+      `Error exporting secrets to files`,
+      'fail'
+    );
+    printError(error);
+  }
+  return false;
+}
