+++ src/ops/MappingOps.ts
@@ -184,65 +184,6 @@ export async function exportMappingsToFiles(
   return false;
 }
 
-function processMappings(mappings, targetDir, name) {
-  try {
-    mappings.forEach((mapping) => {
-      if (name && name !== mapping.name) {
-        return;
-      }
-
-      const mappingPath = `${targetDir}`;
-
-      if (!fs.existsSync(mappingPath)) {
-        fs.mkdirSync(mappingPath, { recursive: true });
-      }
-
-      Object.entries(mapping).forEach(([key, value]) => {
-        if (
-          typeof value === 'object' &&
-          value !== null &&
-          'type' in value &&
-          'source' in value &&
-          value.type === 'text/javascript'
-        ) {
-          const scriptFilename = `${mapping.name}.${key}.js`;
-          (value as any).file = scriptFilename; // Replace source code with file reference
-          fs.writeFileSync(
-            path.join(mappingPath, scriptFilename),
-            (value as any).source
-          );
-          delete (value as any).source;
-        }
-      });
-
-      const fileName = `${mappingPath}/${mapping.name}.json`;
-      saveJsonToFile(mapping, fileName, false);
-    });
-  } catch (err) {
-    console.error(err);
-  }
-}
-
-export async function configManagerExportMappings(
-  options: MappingExportOptions = {
-    deps: true,
-    useStringArrays: true,
-  }
-): Promise<boolean> {
-  try {
-    const exportData = await exportMappings(options);
-    const fileDir = `sync/mappings`;
-    for (const mapping of Object.values(exportData.sync.mappings)) {
-      processMappings([mapping], `${fileDir}/${mapping.name}`, mapping.name);
-    }
-
-    return true;
-  } catch (error) {
-    printError(error, `Error exporting mappings to files`);
-  }
-  return false;
-}
-
 /**
  * Import a mapping from file
  * @param {string} mappingId mapping id/name
